% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_median_ci.R
\name{bootstrap_median_ci}
\alias{bootstrap_median_ci}
\title{Bootstrap confidence interval for the median}
\usage{
bootstrap_median_ci(x, B = 1000, conf = 0.95)
}
\arguments{
\item{x}{Numeric vector of observations.}

\item{B}{Number of bootstrap samples.}

\item{conf}{Confidence level between 0 and 1.}
}
\value{
An S3 object of class "bootstrap_median_ci" containing
the sample median, confidence interval, and number of bootstrap
samples.
}
\description{
Calculates a percentile bootstrap confidence interval for
the population median.
}
